pacstrap: Import all keyrings in /usr/share/pacman/keyrings
With this change, pacman-key will populate the pacman keyring in the
chroot with EVERY keyring found in /usr/share/pacman/keyrings on the
host.

For example, assuming you're building a Manjaro image, you will want to
make sure that the host has both archlinux and manjaro keyrings in
/usr/share/pacman/keyrings. Then running `pacman-key --populate` without
argument will effectively populate the chroot keyring with both.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>